WebServer/requirements.txt
bcrypt==4.1.1
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
colorama==0.4.6
cryptography==41.0.7
deprecation==2.1.0
distlib==0.3.7
filelock==3.13.1
firebase-admin==6.2.0
firebase-functions==0.1.2
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
functions-framework==3.5.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
msgpack==1.0.7
packaging==23.2
pipenv==2023.11.15
platformdirs==3.11.0
proto-plus==1.22.3
protobuf==4.25.1
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
SQLAlchemy==2.0.23
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
waitress==2.1.2
watchdog==3.0.0
Werkzeug==3.0.1
